fix componente i and cupon descuento and solve obligacion remove

+++ pcs/src/main/scala/consumers/registral/componente_i/application/entities/ComponenteIExternalDto.scala
@@ -21,7 +21,7 @@ import serialization.CbroSerialization
                                  distributionId:Option[String],
                                  codigo: Option[String],
                                  concepto: Option[String],
-                                 monto: Option[BigInt]
+                                 monto: Option[BigDecimal]
                                ) extends CbroSerialization
 
 case class ListDetallesComponenteI(BCI_DETALLES: List[DetallesComponenteI]) extends CbroSerialization
